Add files via upload
Added code to compute local and global mass matrix. Also added rudimentary version of versem.py, where further refining is needed in terms of the input variables. Currently a lot of stuff is hardcoded. This should instead be passed on from the input file defined by the user or channelized from the mesh properties.